Removed code duplication in update stack step and create change set step.
